% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fdr.R
\name{cluster-fdr}
\alias{cluster-fdr}
\title{Cluster-Level FDR Control}
\description{
Implements topological FDR control at the cluster level using
Benjamini-Hochberg procedure applied to cluster p-values.
}
