% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfce.R
\name{.array_to_vol}
\alias{.array_to_vol}
\title{Convert array to vol-like object (placeholder)}
\usage{
.array_to_vol(vals, dims, template)
}
\description{
Convert array to vol-like object (placeholder)
}
\keyword{internal}
