% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fdr.R
\name{.cluster_fdr_twosided}
\alias{.cluster_fdr_twosided}
\title{Two-sided cluster FDR}
\usage{
.cluster_fdr_twosided(
  stat_vol,
  mask,
  cluster_thresh,
  q,
  p_method,
  fwhm_mm,
  n_perm,
  two_sided_policy,
  seed
)
}
\description{
Two-sided cluster FDR
}
\keyword{internal}
