% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_fdr.R
\name{.cluster_pvals_perm}
\alias{.cluster_pvals_perm}
\title{Permutation cluster p-values}
\usage{
.cluster_pvals_perm(z, mask, u0, obs_clusters, n_perm, seed)
}
\description{
Permutation cluster p-values
}
\keyword{internal}
