% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rft.R
\name{.find_clusters}
\alias{.find_clusters}
\title{Find connected components (simple 6-connectivity)}
\usage{
.find_clusters(binary_mask, dims)
}
\description{
Find connected components (simple 6-connectivity)
}
\keyword{internal}
