% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{.score_parcels_vec}
\alias{.score_parcels_vec}
\title{Vectorized parcel scoring}
\usage{
.score_parcels_vec(z_vec, pi_vec, lab_vec, log_den1, inv_sqrt_den2, kappa_grid)
}
\description{
Vectorized parcel scoring
}
\keyword{internal}
