% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfce.R
\name{.tfce_from_vec}
\alias{.tfce_from_vec}
\title{TFCE from vector (internal helper)}
\usage{
.tfce_from_vec(z_vec, dims, mask, H, E, dh, tail)
}
\description{
TFCE from vector (internal helper)
}
\keyword{internal}
