% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfce.R
\name{.tfce_fwer_twosided}
\alias{.tfce_fwer_twosided}
\title{Two-sided TFCE with combined max null}
\usage{
.tfce_fwer_twosided(stat_vol, mask, n_perm, H, E, dh, alpha, seed)
}
\description{
Two-sided TFCE with combined max null
}
\keyword{internal}
