% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_scan.R
\name{print.neurothresh_result}
\alias{print.neurothresh_result}
\title{Print method for neurothresh_result}
\usage{
\method{print}{neurothresh_result}(x, ...)
}
\arguments{
\item{x}{A neurothresh_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for neurothresh_result
}
