% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rft.R
\name{rft-methods}
\alias{rft-methods}
\title{Random Field Theory (RFT) Thresholding Methods}
\description{
Classic GRF/RFT inference using expected Euler characteristic
for FWER control. These are provided as baseline comparisons
for LR-MFT.
}
