% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rft.R
\name{rft_fwer_twosided}
\alias{rft_fwer_twosided}
\title{Two-sided RFT wrapper}
\usage{
rft_fwer_twosided(stat_vol, method = c("peak", "cluster"), ...)
}
\arguments{
\item{stat_vol}{NeuroVol containing statistic map}

\item{method}{"peak" or "cluster"}

\item{...}{Additional arguments passed to rft_peak_fwer or rft_cluster_fwer}
}
\description{
Two-sided RFT wrapper
}
