% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octree.R
\name{split_indices_bbox}
\alias{split_indices_bbox}
\title{Split indices with bounding boxes}
\usage{
split_indices_bbox(idx_parent, split, x, y, z)
}
\arguments{
\item{idx_parent}{Integer vector of parent mask-space indices}

\item{split}{List returned by \code{\link{octree_split_info}}}

\item{x}{Integer vector of x coordinates}

\item{y}{Integer vector of y coordinates}

\item{z}{Integer vector of z coordinates}
}
\value{
List with components:
\describe{
\item{idx}{List of m IntegerVectors with child indices}
\item{bbox}{Integer matrix (m x 6) with child bounding boxes}
}
}
\description{
Splits parent indices into children and computes bounding box
for each child in a single pass.
}
