% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_scan.R
\name{summary.neurothresh_result}
\alias{summary.neurothresh_result}
\title{Summary method for neurothresh_result}
\usage{
\method{summary}{neurothresh_result}(object, ...)
}
\arguments{
\item{object}{A neurothresh_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Summary method for neurothresh_result
}
