% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfce.R
\name{tfce-methods}
\alias{tfce-methods}
\title{Threshold-Free Cluster Enhancement (TFCE)}
\description{
Implementation of FSL-style TFCE for threshold-free cluster
enhancement with permutation-based FWER control.
}
